// generated/backend/models/TextFromAudioWithCallInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { NoteDetailsCompletion } from './NoteDetailsCompletion';
export type TextFromAudioWithCallInfo = {
    completionFromAudio?: NoteDetailsCompletion;
    rawSRT?: string;
    endTimestamp?: string;
};